Fix import and TypeScript types to be interface or type
