+++ src/components/common/types.ts
@@ -1,0 +1,11 @@
+type Page = {
+  name: string;
+  path: string;
+};
+
+type KeyValuePair<K, V> = {
+  key: K;
+  value: V;
+};
+
+export type { Page, KeyValuePair };
