+++ src/components/products/models.ts
@@ -1,0 +1,13 @@
+interface Product {
+  id: number;
+  name: string;
+  price: number;
+  unit: string;
+  description: string;
+}
+
+interface ProductPhoto {
+  url: string;
+}
+
+export type { Product, ProductPhoto };
