+++ src/constants/appConfig.ts
@@ -1,2 +1,4 @@
-export const APPLICATION_NAME: string = import.meta.env.VITE_APP_NAME;
-export const BACKEND_URL: string = import.meta.env.VITE_BACKEND_URL;
+const APPLICATION_NAME: string = import.meta.env.VITE_APP_NAME;
+const BACKEND_URL: string = import.meta.env.VITE_BACKEND_URL;
+
+export { APPLICATION_NAME, BACKEND_URL };
