+++ src/services/productPhotoService.ts
@@ -1,6 +1,7 @@
 import axiosInstance from "axios";
 
-import { BACKEND_URL } from "../constants/appConfig";
+import type { ProductPhoto } from "../components/products/models.ts";
+import { BACKEND_URL } from "../constants/appConfig.ts";
 
 const getProductPhotos = async (productId: number): Promise<ProductPhoto[]> => {
   const response = await axiosInstance.get<ProductPhoto[]>(
